% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{matlab.like.colorRamps}
\alias{matlab.like.colorRamps}
\title{Generate color palettes similar to the Matlab default}
\usage{
matlab.like.colorRamps(n, two = FALSE)
}
\arguments{
\item{n}{number of colors in the palette.}

\item{two}{flag indicating whether to use \code{colorRamps::matlab.like} or
\code{colorRamps::matlab.like2}.}
}
\value{
A vector of \code{n} colors.
}
\description{
Generates Matlab-like color palettes. Functions imported from
the colorRamps package.
}
\examples{
image(matrix(1:100, 10), col = matlab.like.colorRamps(100))
image(matrix(1:100, 10), col = matlab.like.colorRamps(100, two = TRUE))
}
\keyword{internal}
