% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{plotVmdSurface3D}
\alias{plotVmdSurface3D}
\title{Visualization of a 3D surface using VMD}
\usage{
plotVmdSurface3D(t, nx, ny, nz, outFile = "test.map", call = FALSE, path)
}
\arguments{
\item{t}{a vector of length \code{nx * ny * nz} containing the evaluation of the function to be represented, i.e. \code{t = f(expand.grid(x, y, z))}.}

\item{nx, ny, nz}{length of the grid along each dimension.}

\item{outFile}{file in XPLOR format to be read by VMD.}

\item{call}{whether to invoke VMD for immediate visualization or not.}

\item{path}{path to VMD in the system. See details.}
}
\value{
The function creates a XPLOR map file in the current working directory. This file contains \code{t} and other data that allows its readily interpretation by VMD. An optional call to VMD can also be done for immediate visualization.
}
\description{
Visualization of a real function of three variables by means of the 3D contour feature of VMD. The function creates an XPLOR map file that can be read by VMD and, optionally, establishes a connection for inspection of the surface.
}
\details{
If \code{call = TRUE}, then a call to VMD with \code{outFile} is produced. VMD needs to be installed (\url{http://www.ks.uiuc.edu/Research/vmd/}) and its \code{path} must be properly specified. For example, for Mac OS it should be something similar to \code{path = "/Applications/VMD\\ 1.9.2.app/Contents/MacOs/startup.command"}, with the version number depending on the installed one. The R session will wait until finalization of VMD.

After launching VMD, a 3D scatterplot is shown. The options for exploration of the 3D contours can be accessed in "VMD Main -> Graphics -> Representations ...". This will open a new window with the available choices. The drawing methods 'Isosurface', 'VolumeSlice' and 'Fieldlines' are the most useful.

Recall that the spacing is assumed to be the same in each of the grids.

See \url{http://www.ks.uiuc.edu/Research/vmd/} for help on using VMD.
}
\examples{
\dontrun{
gridx <- seq(-pi, pi, l = 50)
gridy <- seq(0, pi, l = 50)
gridz <- seq(0, pi, l = 20)
t <- plotSurface3D(gridx, gridy, gridz, size = 10, alpha = 0.01, fVect = TRUE,
                   f = function(x) 10 * (sin(x[, 1]) * cos(x[, 2]) - sin(x[, 3]))^2)

# VMD has to be installed in the system (http://www.ks.uiuc.edu/Research/vmd/)
# Also, the path must be properly specified
plotVmdSurface3D(t = t, nx = length(gridx), ny = length(gridy), nz = length(gridz),
                 outFile = "test.map", call = TRUE, path =
                 "/Applications/VMD\\\\ 1.9.2.app/Contents/MacOs/startup.command")
}
}
\references{
Humphrey, W., Dalke, A. and Schulten, K. (1996). VMD - Visual Molecular Dynamics, \emph{J. Molec. Graphics}, 14(1):33-38.
}
