% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{repRow}
\alias{repRow}
\alias{repCol}
\title{Replication of rows and columns}
\usage{
repRow(x, n)

repCol(x, n)
}
\arguments{
\item{x}{a numerical vector or matrix of dimension \code{c(nr, nc)}.}

\item{n}{the number of replicates of \code{x} by rows or columns.}
}
\value{
A matrix of dimension \code{c(nr * n, nc)} for \code{repRow} or
\code{c(nr, nc * n)} for \code{repCol}.
}
\description{
Wrapper for replicating a matrix/vector by rows or columns.
}
\examples{
repRow(1:5, 2)
repCol(1:5, 2)
A <- rbind(1:5, 5:1)
A
repRow(A, 2)
repCol(A, 2)
}
\keyword{internal}
