% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{table.ramp.colorRamps}
\alias{table.ramp.colorRamps}
\title{Constructs color palettes with sharp breaks}
\usage{
table.ramp.colorRamps(n, mid = 0.5, sill = 0.5, base = 1, height = 1)
}
\description{
See \code{?colorRamps::rgb.tables}.
}
\keyword{internal}
